feat: implement progress tracking and CLI execution flow
Add interfaces and implementations for progress tracking during image optimization. Introduce a command-line interface (CLI) executor to handle user input and execute optimization tasks. Remove obsolete file utility code to streamline functionality.
