+++ src/optimus-image/cmd/executer.go
@@ -1,0 +1,6 @@
+package cmd
+
+// Executor defines the contract for CLI execution.
+type Executor interface {
+	Execute(getSelectionFunc func() (string, error)) (string, error)
+}
