+++ src/optimus-image/cmd/root.go
@@ -3,34 +3,29 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/t0nylombardi/optimus-image/src/optimus-image/internal/file"
 	"github.com/t0nylombardi/optimus-image/src/optimus-image/internal/optimizer"
-	"github.com/t0nylombardi/optimus-image/src/optimus-image/internal/utils"
-
-	"github.com/manifoldco/promptui"
+	"github.com/t0nylombardi/optimus-image/src/optimus-image/internal/progress"
 )
 
-type Executor struct {
-	FileUtils     utils.FileUtils
+// ExecutorImpl handles CLI execution.
+type ExecutorImpl struct {
+	FileUtils     file.FileUtils
 	FileOptimizer optimizer.FileOptimizer
+	Tracker       progress.ProgressTracker
 }
 
-// GetUserSelection prompts the user to choose between processing a single file or a directory.
-func GetUserSelection() (string, error) {
-	prompt := promptui.Select{
-		Label: "Choose input type",
-		Items: []string{"Single File", "Directory"},
+// NewExecutor creates a new ExecutorImpl instance.
+func NewExecutor(fileUtils file.FileUtils, fileOptimizer optimizer.FileOptimizer, tracker progress.ProgressTracker) Executor {
+	return &ExecutorImpl{
+		FileUtils:     fileUtils,
+		FileOptimizer: fileOptimizer,
+		Tracker:       tracker,
 	}
-
-	_, result, err := prompt.Run()
-	if err != nil {
-		return "", fmt.Errorf("prompt failed: %v", err)
-	}
-	return result, nil
 }
 
 // Execute runs the CLI workflow based on user selection.
-func (e *Executor) Execute(getSelectionFunc func() (string, error)) (string, error) {
-	// Get user input for input type (Single File / Directory)
+func (e *ExecutorImpl) Execute(getSelectionFunc func() (string, error)) (string, error) {
 	selection, err := getSelectionFunc()
 	if err != nil {
 		return "", err
@@ -48,42 +43,38 @@ func (e *Executor) Execute(getSelectionFunc func() (string, error)) (string, err
 	}
 }
 
-// processSingleFile handles the workflow for optimizing a single image file.
-func (e *Executor) processSingleFile() error {
-	// Get file path from the user
+// processSingleFile optimizes a single image file.
+func (e *ExecutorImpl) processSingleFile() error {
 	filePath, err := e.FileUtils.GetFilePath()
 	if err != nil {
 		return err
 	}
 
-	// Run optimization process
 	files := []string{filePath}
-
-	if err := e.FileOptimizer.OptimizeFiles(files); err != nil {
+	if err := e.FileOptimizer.OptimizeFiles(files, e.Tracker); err != nil {
 		return err
 	}
 
-	fmt.Println("Image optimization complete!")
+	fmt.Println("✅ Image optimization complete!")
 	return nil
 }
 
-// processDirectory handles the workflow for optimizing all image files in a directory.
-func (e *Executor) processDirectory() error {
-	// Get directory path from the user
+// processDirectory optimizes all images in a directory.
+func (e *ExecutorImpl) processDirectory() error {
 	dirPath, err := e.FileUtils.GetDirectoryPath()
 	if err != nil {
 		return err
 	}
 
-	// Run optimization process
 	files, err := e.FileUtils.GetFilesInDirectory(dirPath)
 	if err != nil {
 		return err
 	}
 
-	if err := e.FileOptimizer.OptimizeFiles(files); err != nil {
+	if err := e.FileOptimizer.OptimizeFiles(files, e.Tracker); err != nil {
 		return err
 	}
 
+	fmt.Println("✅ Directory optimization complete!")
 	return nil
 }
