+++ src/optimus-image/internal/optimizer/optimize_test.go
@@ -1,0 +1,1 @@
+package optimizer
