+++ src/optimus-image/internal/optimizer/optimizer.go
@@ -1,0 +1,8 @@
+package optimizer
+
+import "github.com/t0nylombardi/optimus-image/src/optimus-image/internal/progress"
+
+// FileOptimizer defines the contract for optimizing images.
+type FileOptimizer interface {
+	OptimizeFiles(images []string, tracker progress.ProgressTracker) error
+}
