+++ src/optimus-image/internal/progress/tracker_test.go
@@ -1,0 +1,1 @@
+package progress
